feat: refactor карточки
